DATAGRAPH-1149 - Introduce MetaDataProvider.
The newly introduced MetaDataProvider is used to retrieve the OGM MetaData from the session factory as the „source of truth.“ Getting them from the session itself is hacky and can lead to the creation of a huge amount of unnessary transactions.

As SDN works only on the session object and keeps the session factory hidden, we faciliate the shared session creator to add this interface to the session and delegate back to the session factory. If the proxy isn’t used, we’ll try to lookup the meta data on the concrete Neo4jSession.

Two test that mock the session and don’t use the proxy had to be fixed to include the additional interface.